+++ CHANGELOG.md
@@ -45,7 +45,7 @@ and this project somewhat adheres to [Semantic Versioning](https://semver.org/sp
 - Fix Undo/Redo ignoring the automatic resizing that occurs if a layout/border was an unexpected size.
 - Fix Undo/Redo in the Tileset and Palette Editors and Paste in the Tileset Editor appearing enabled even when they don't do anything.
 - Fix `Ctrl+Shift+Z` not being set as a default shortcut for Redo in the Palette Editor like it is for other windows.
-- Fix the Tileset Editor's status bar not updating while selecting tiles in the metatile layer view.
+- Fix the Tileset Editor's status bar not updating while selecting tiles in the metatile layer view, or when pasting metatiles.
 - Fix the main window's status bar not immediately reflecting changes made while painting metatiles / movement permissions.
 - Fix cleared metatile labels not updating until the project is reloaded.
 - Fix some changes in the Tileset Editor being discarded if the window is closed too quickly.